add comment about automatic awecron repo detection
This feature is very useful for portability, however I am not a fan of
it because I feel like something might break, and considering that
awecron runs as root this would be horrible. In addition to that, in the
future the awecron executable probably will not be in the awecron repo,
so this will be useless. Nevertheless, for now I will still keep it with
this comment.