add error run feature
Recently I had some issues with one of my cronjobs that need internet
connection in order to work. The problem was that it seems like awecron
sometimes started before internet connection was established. As the
result of this it exits with exit code 0, but it was not actually
successful. This simple feature is supposed to allow setting it to
exit with a special exit code that will re-run the cronjob after smaller
interval. The feature is not standard, but it is still fairly minimal.