revert some changes and add parallelism as experimental
It was very tempting to try implementing parallelism for cronjobs and I
managed to do it relatively easy, however the actual challange was to
account for the environment. When running a lot of cronjobs at the same
time the resource usage may result in very high spikes that might be
even noticable. To prevent that I was trying to create a throttling
system based on CPU and RAM usage percentage. I managed to easily
implement this simple idea at the base of `main` function, but this
means that it will run on all cronjobs including those that won't need
to run which is not optimized enough for me. The solution for this would
be to put throttling into the cronjob run function, but the tldr problem
is that it is separate environment (because of the parantheses) and so I
can not control the background processes across these environments
(which is the problem). I can not remove the parantheses because it is
one of the main error prevention features. My final solution was to
simply forget about throttling and call it an "experimental" feature,
and to enable it the user will have to uncomment some lines in awecron.