add cronjob timeout as experimental
This is another important error handling feature that I really wanted to
implement. The simplest solution would be to just use GNU timeout, but
it is not optimal because it is GNU specific and I don't want another
dependency. After a research and testing I ended up with an optimal
cross platform bash code that supposed to work but in reality it does
not and even breaks awecron. After a lot of testing I still can not
figure out what is the problem. As a temporary solution I have
implemented timeout with GNU timeout as experimental feature, but this
should be fixed in the future. The following is that simple code that
supposed to work but doesn't (put it where the run function is called):

set +b
run &
sleep 10 &
wait -n
kill -SIGKILL $(jobs -p)