fix error handling and improve timeout functionality
I have noticed a problem with error handling after the previous commit.
The problem was that it is difficult to differentiate when the cronjob
exits and when it times out to know if it is necessary to stop the
watchdog process. The chosen solution was to blindly try to stop the
watchdog process. It is not ideal, however implementing more complex
logic is not ideal either. Another improvement was to force quit the
cronjob instead of sending TERM signal which could be ignored.